carve out token v2 processor (#83)
