+++ rust/processor/src/models/default_models/v2_objects.rs
@@ -7,9 +7,9 @@
 
 use super::move_resources::MoveResource;
 use crate::{
-    models::token_models::{
-        collection_datas::{QUERY_RETRIES, QUERY_RETRY_DELAY_MS},
-        v2_token_utils::ObjectWithMetadata,
+    models::{
+        token_models::collection_datas::{QUERY_RETRIES, QUERY_RETRY_DELAY_MS},
+        token_v2_models::v2_token_utils::ObjectWithMetadata,
     },
     schema::{current_objects, objects},
     utils::database::PgPoolConnection,
