+++ rust/processor/src/models/fungible_asset_models/v2_fungible_asset_balances.rs
@@ -13,7 +13,7 @@ use super::{
 use crate::{
     models::{
         coin_models::coin_utils::{CoinInfoType, CoinResource},
-        token_models::v2_token_utils::TokenStandard,
+        token_v2_models::v2_token_utils::TokenStandard,
     },
     schema::{current_fungible_asset_balances, fungible_asset_balances},
     utils::{database::PgPoolConnection, util::standardize_address},
