+++ rust/processor/src/models/mod.rs
@@ -9,3 +9,4 @@ pub mod processor_status;
 pub mod property_map;
 pub mod stake_models;
 pub mod token_models;
+pub mod token_v2_models;
