+++ rust/processor/src/models/token_models/mod.rs
@@ -10,9 +10,3 @@ pub mod token_datas;
 pub mod token_ownerships;
 pub mod token_utils;
 pub mod tokens;
-pub mod v2_collections;
-pub mod v2_token_activities;
-pub mod v2_token_datas;
-pub mod v2_token_metadata;
-pub mod v2_token_ownerships;
-pub mod v2_token_utils;
