+++ rust/processor/src/models/token_v2_models/mod.rs
@@ -1,0 +1,9 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod v2_collections;
+pub mod v2_token_activities;
+pub mod v2_token_datas;
+pub mod v2_token_metadata;
+pub mod v2_token_ownerships;
+pub mod v2_token_utils;
