+++ rust/processor/src/models/token_v2_models/v2_token_utils.rs
@@ -4,7 +4,6 @@
 // This is required because a diesel macro makes clippy sad
 #![allow(clippy::extra_unused_lifetimes)]
 
-use super::token_utils::{NAME_LENGTH, URI_LENGTH};
 use crate::{
     models::{
         coin_models::coin_utils::COIN_ADDR,
@@ -12,6 +11,7 @@ use crate::{
         fungible_asset_models::v2_fungible_asset_utils::{
             FungibleAssetMetadata, FungibleAssetStore, FungibleAssetSupply,
         },
+        token_models::token_utils::{NAME_LENGTH, URI_LENGTH},
     },
     utils::util::{
         deserialize_from_string, deserialize_token_object_property_map_from_bcs_hexstring,
