+++ rust/processor/src/processors/mod.rs
@@ -7,11 +7,13 @@ pub mod fungible_asset_processor;
 pub mod processor_trait;
 pub mod stake_processor;
 pub mod token_processor;
+pub mod token_v2_processor;
 
 use self::{
     coin_processor::NAME as COIN_PROCESSOR_NAME, default_processor::NAME as DEFAULT_PROCESSOR_NAME,
     fungible_asset_processor::NAME as FUNGIBLE_ASSET_PROCESSOR_NAME,
     stake_processor::NAME as STAKE_PROCESSOR_NAME, token_processor::NAME as TOKEN_PROCESSOR_NAME,
+    token_v2_processor::NAME as TOKEN_V2_PROCESSOR_NAME,
 };
 
 pub enum Processor {
@@ -20,6 +22,7 @@ pub enum Processor {
     FungibleAssetProcessor,
     StakeProcessor,
     TokenProcessor,
+    TokenV2Processor,
 }
 
 impl Processor {
@@ -30,6 +33,7 @@ impl Processor {
             FUNGIBLE_ASSET_PROCESSOR_NAME => Self::FungibleAssetProcessor,
             STAKE_PROCESSOR_NAME => Self::StakeProcessor,
             TOKEN_PROCESSOR_NAME => Self::TokenProcessor,
+            TOKEN_V2_PROCESSOR_NAME => Self::TokenV2Processor,
             _ => panic!("Processor unsupported {}", input_str),
         }
     }
