+++ rust/processor/src/processors/token_processor.rs
@@ -3,56 +3,30 @@
 
 use super::processor_trait::{ProcessingResult, ProcessorTrait};
 use crate::{
-    models::{
-        fungible_asset_models::v2_fungible_asset_utils::{
-            FungibleAssetMetadata, FungibleAssetStore, FungibleAssetSupply,
-        },
-        token_models::{
-            ans_lookup::{CurrentAnsLookup, CurrentAnsLookupPK},
-            collection_datas::{CollectionData, CurrentCollectionData},
-            nft_points::NftPoints,
-            token_activities::TokenActivity,
-            token_claims::CurrentTokenPendingClaim,
-            token_datas::{CurrentTokenData, TokenData},
-            token_ownerships::{CurrentTokenOwnership, TokenOwnership},
-            tokens::{
-                CurrentTokenOwnershipPK, CurrentTokenPendingClaimPK, TableHandleToOwner,
-                TableMetadataForToken, Token, TokenDataIdHash,
-            },
-            v2_collections::{CollectionV2, CurrentCollectionV2, CurrentCollectionV2PK},
-            v2_token_activities::TokenActivityV2,
-            v2_token_datas::{CurrentTokenDataV2, CurrentTokenDataV2PK, TokenDataV2},
-            v2_token_metadata::{CurrentTokenV2Metadata, CurrentTokenV2MetadataPK},
-            v2_token_ownerships::{
-                CurrentTokenOwnershipV2, CurrentTokenOwnershipV2PK, NFTOwnershipV2,
-                TokenOwnershipV2,
-            },
-            v2_token_utils::{
-                AptosCollection, BurnEvent, FixedSupply, ObjectWithMetadata, PropertyMapModel,
-                TokenV2, TokenV2AggregatedData, TokenV2AggregatedDataMapping, TokenV2Burned,
-                TransferEvent, UnlimitedSupply,
-            },
+    models::token_models::{
+        ans_lookup::{CurrentAnsLookup, CurrentAnsLookupPK},
+        collection_datas::{CollectionData, CurrentCollectionData},
+        nft_points::NftPoints,
+        token_activities::TokenActivity,
+        token_claims::CurrentTokenPendingClaim,
+        token_datas::{CurrentTokenData, TokenData},
+        token_ownerships::{CurrentTokenOwnership, TokenOwnership},
+        tokens::{
+            CurrentTokenOwnershipPK, CurrentTokenPendingClaimPK, TableMetadataForToken, Token,
+            TokenDataIdHash,
         },
     },
     schema,
-    utils::{
-        database::{
-            clean_data_for_db, execute_with_better_error, get_chunks, PgDbPool, PgPoolConnection,
-        },
-        util::{get_entry_function_from_user_request, parse_timestamp, standardize_address},
+    utils::database::{
+        clean_data_for_db, execute_with_better_error, get_chunks, PgDbPool, PgPoolConnection,
     },
 };
 use anyhow::bail;
-use aptos_indexer_protos::transaction::v1::{
-    transaction::TxnData, write_set_change::Change, Transaction,
-};
+use aptos_indexer_protos::transaction::v1::Transaction;
 use async_trait::async_trait;
 use diesel::{pg::upsert::excluded, result::Error, ExpressionMethods, PgConnection};
 use field_count::FieldCount;
-use std::{
-    collections::{HashMap, HashSet},
-    fmt::Debug,
-};
+use std::{collections::HashMap, fmt::Debug};
 use tracing::error;
 
 pub const NAME: &str = "token_processor";
@@ -103,25 +77,6 @@ fn insert_to_db_impl(
     current_token_claims: &[CurrentTokenPendingClaim],
     current_ans_lookups: &[CurrentAnsLookup],
     nft_points: &[NftPoints],
-    (
-        collections_v2,
-        token_datas_v2,
-        token_ownerships_v2,
-        current_collections_v2,
-        current_token_datas_v2,
-        current_token_ownerships_v2,
-        token_activities_v2,
-        current_token_v2_metadata,
-    ): (
-        &[CollectionV2],
-        &[TokenDataV2],
-        &[TokenOwnershipV2],
-        &[CurrentCollectionV2],
-        &[CurrentTokenDataV2],
-        &[CurrentTokenOwnershipV2],
-        &[TokenActivityV2],
-        &[CurrentTokenV2Metadata],
-    ),
 ) -> Result<(), diesel::result::Error> {
     let (tokens, token_ownerships, token_datas, collection_datas) = basic_token_transaction_lists;
     let (current_token_ownerships, current_token_datas, current_collection_datas) =
@@ -137,14 +92,6 @@ fn insert_to_db_impl(
     insert_current_token_claims(conn, current_token_claims)?;
     insert_current_ans_lookups(conn, current_ans_lookups)?;
     insert_nft_points(conn, nft_points)?;
-    insert_collections_v2(conn, collections_v2)?;
-    insert_token_datas_v2(conn, token_datas_v2)?;
-    insert_token_ownerships_v2(conn, token_ownerships_v2)?;
-    insert_current_collections_v2(conn, current_collections_v2)?;
-    insert_current_token_datas_v2(conn, current_token_datas_v2)?;
-    insert_current_token_ownerships_v2(conn, current_token_ownerships_v2)?;
-    insert_token_activities_v2(conn, token_activities_v2)?;
-    insert_current_token_v2_metadatas(conn, current_token_v2_metadata)?;
     Ok(())
 }
 
@@ -168,25 +115,6 @@ fn insert_to_db(
     current_token_claims: Vec<CurrentTokenPendingClaim>,
     current_ans_lookups: Vec<CurrentAnsLookup>,
     nft_points: Vec<NftPoints>,
-    (
-        collections_v2,
-        token_datas_v2,
-        token_ownerships_v2,
-        current_collections_v2,
-        current_token_datas_v2,
-        current_token_ownerships_v2,
-        token_activities_v2,
-        current_token_v2_metadata,
-    ): (
-        Vec<CollectionV2>,
-        Vec<TokenDataV2>,
-        Vec<TokenOwnershipV2>,
-        Vec<CurrentCollectionV2>,
-        Vec<CurrentTokenDataV2>,
-        Vec<CurrentTokenOwnershipV2>,
-        Vec<TokenActivityV2>,
-        Vec<CurrentTokenV2Metadata>,
-    ),
 ) -> Result<(), diesel::result::Error> {
     tracing::trace!(
         name = name,
@@ -213,16 +141,6 @@ fn insert_to_db(
                 &current_token_claims,
                 &current_ans_lookups,
                 &nft_points,
-                (
-                    &collections_v2,
-                    &token_datas_v2,
-                    &token_ownerships_v2,
-                    &current_collections_v2,
-                    &current_token_datas_v2,
-                    &current_token_ownerships_v2,
-                    &token_activities_v2,
-                    &current_token_v2_metadata,
-                ),
             )
         }) {
         Ok(_) => Ok(()),
@@ -241,15 +159,6 @@ fn insert_to_db(
                 let current_token_claims = clean_data_for_db(current_token_claims, true);
                 let current_ans_lookups = clean_data_for_db(current_ans_lookups, true);
                 let nft_points = clean_data_for_db(nft_points, true);
-                let collections_v2 = clean_data_for_db(collections_v2, true);
-                let token_datas_v2 = clean_data_for_db(token_datas_v2, true);
-                let token_ownerships_v2 = clean_data_for_db(token_ownerships_v2, true);
-                let current_collections_v2 = clean_data_for_db(current_collections_v2, true);
-                let current_token_datas_v2 = clean_data_for_db(current_token_datas_v2, true);
-                let current_token_ownerships_v2 =
-                    clean_data_for_db(current_token_ownerships_v2, true);
-                let token_activities_v2 = clean_data_for_db(token_activities_v2, true);
-                let current_token_v2_metadata = clean_data_for_db(current_token_v2_metadata, true);
 
                 insert_to_db_impl(
                     pg_conn,
@@ -263,16 +172,6 @@ fn insert_to_db(
                     &current_token_claims,
                     &current_ans_lookups,
                     &nft_points,
-                    (
-                        &collections_v2,
-                        &token_datas_v2,
-                        &token_ownerships_v2,
-                        &current_collections_v2,
-                        &current_token_datas_v2,
-                        &current_token_ownerships_v2,
-                        &token_activities_v2,
-                        &current_token_v2_metadata,
-                    ),
                 )
             }),
     }
@@ -591,231 +490,6 @@ fn insert_nft_points(
     Ok(())
 }
 
-fn insert_collections_v2(
-    conn: &mut PgConnection,
-    items_to_insert: &[CollectionV2],
-) -> Result<(), diesel::result::Error> {
-    use schema::collections_v2::dsl::*;
-
-    let chunks = get_chunks(items_to_insert.len(), CollectionV2::field_count());
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::collections_v2::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict((transaction_version, write_set_change_index))
-                .do_nothing(),
-            None,
-        )?;
-    }
-    Ok(())
-}
-
-fn insert_token_datas_v2(
-    conn: &mut PgConnection,
-    items_to_insert: &[TokenDataV2],
-) -> Result<(), diesel::result::Error> {
-    use schema::token_datas_v2::dsl::*;
-
-    let chunks = get_chunks(items_to_insert.len(), TokenDataV2::field_count());
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::token_datas_v2::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict((transaction_version, write_set_change_index))
-                .do_nothing(),
-            None,
-        )?;
-    }
-    Ok(())
-}
-
-fn insert_token_ownerships_v2(
-    conn: &mut PgConnection,
-    items_to_insert: &[TokenOwnershipV2],
-) -> Result<(), diesel::result::Error> {
-    use schema::token_ownerships_v2::dsl::*;
-
-    let chunks = get_chunks(items_to_insert.len(), TokenOwnershipV2::field_count());
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::token_ownerships_v2::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict((transaction_version, write_set_change_index))
-                .do_nothing(),
-            None,
-        )?;
-    }
-    Ok(())
-}
-
-fn insert_current_collections_v2(
-    conn: &mut PgConnection,
-    items_to_insert: &[CurrentCollectionV2],
-) -> Result<(), diesel::result::Error> {
-    use schema::current_collections_v2::dsl::*;
-
-    let chunks = get_chunks(items_to_insert.len(), CurrentCollectionV2::field_count());
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::current_collections_v2::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict(collection_id)
-                .do_update()
-                .set((
-                    creator_address.eq(excluded(creator_address)),
-                    collection_name.eq(excluded(collection_name)),
-                    description.eq(excluded(description)),
-                    uri.eq(excluded(uri)),
-                    current_supply.eq(excluded(current_supply)),
-                    max_supply.eq(excluded(max_supply)),
-                    total_minted_v2.eq(excluded(total_minted_v2)),
-                    mutable_description.eq(excluded(mutable_description)),
-                    mutable_uri.eq(excluded(mutable_uri)),
-                    table_handle_v1.eq(excluded(table_handle_v1)),
-                    token_standard.eq(excluded(token_standard)),
-                    last_transaction_version.eq(excluded(last_transaction_version)),
-                    last_transaction_timestamp.eq(excluded(last_transaction_timestamp)),
-                    inserted_at.eq(excluded(inserted_at)),
-                )),
-            Some(" WHERE current_collections_v2.last_transaction_version <= excluded.last_transaction_version "),
-        )?;
-    }
-    Ok(())
-}
-
-fn insert_current_token_datas_v2(
-    conn: &mut PgConnection,
-    items_to_insert: &[CurrentTokenDataV2],
-) -> Result<(), diesel::result::Error> {
-    use schema::current_token_datas_v2::dsl::*;
-
-    let chunks = get_chunks(items_to_insert.len(), CurrentTokenDataV2::field_count());
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::current_token_datas_v2::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict(token_data_id)
-                .do_update()
-                .set((
-                    collection_id.eq(excluded(collection_id)),
-                    token_name.eq(excluded(token_name)),
-                    maximum.eq(excluded(maximum)),
-                    supply.eq(excluded(supply)),
-                    largest_property_version_v1.eq(excluded(largest_property_version_v1)),
-                    token_uri.eq(excluded(token_uri)),
-                    description.eq(excluded(description)),
-                    token_properties.eq(excluded(token_properties)),
-                    token_standard.eq(excluded(token_standard)),
-                    is_fungible_v2.eq(excluded(is_fungible_v2)),
-                    last_transaction_version.eq(excluded(last_transaction_version)),
-                    last_transaction_timestamp.eq(excluded(last_transaction_timestamp)),
-                    inserted_at.eq(excluded(inserted_at)),
-                    decimals.eq(excluded(decimals)),
-                )),
-            Some(" WHERE current_token_datas_v2.last_transaction_version <= excluded.last_transaction_version "),
-        )?;
-    }
-    Ok(())
-}
-
-fn insert_current_token_ownerships_v2(
-    conn: &mut PgConnection,
-    items_to_insert: &[CurrentTokenOwnershipV2],
-) -> Result<(), diesel::result::Error> {
-    use schema::current_token_ownerships_v2::dsl::*;
-
-    let chunks = get_chunks(
-        items_to_insert.len(),
-        CurrentTokenOwnershipV2::field_count(),
-    );
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::current_token_ownerships_v2::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict((token_data_id, property_version_v1, owner_address, storage_id))
-                .do_update()
-                .set((
-                    amount.eq(excluded(amount)),
-                    table_type_v1.eq(excluded(table_type_v1)),
-                    token_properties_mutated_v1.eq(excluded(token_properties_mutated_v1)),
-                    is_soulbound_v2.eq(excluded(is_soulbound_v2)),
-                    token_standard.eq(excluded(token_standard)),
-                    is_fungible_v2.eq(excluded(is_fungible_v2)),
-                    last_transaction_version.eq(excluded(last_transaction_version)),
-                    last_transaction_timestamp.eq(excluded(last_transaction_timestamp)),
-                    inserted_at.eq(excluded(inserted_at)),
-                    non_transferrable_by_owner.eq(excluded(non_transferrable_by_owner)),
-                )),
-            Some(" WHERE current_token_ownerships_v2.last_transaction_version <= excluded.last_transaction_version "),
-        )?;
-    }
-    Ok(())
-}
-
-fn insert_token_activities_v2(
-    conn: &mut PgConnection,
-    items_to_insert: &[TokenActivityV2],
-) -> Result<(), diesel::result::Error> {
-    use schema::token_activities_v2::dsl::*;
-
-    let chunks = get_chunks(items_to_insert.len(), TokenActivityV2::field_count());
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::token_activities_v2::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict((transaction_version, event_index))
-                .do_update()
-                .set((
-                    entry_function_id_str.eq(excluded(entry_function_id_str)),
-                    inserted_at.eq(excluded(inserted_at)),
-                )),
-            None,
-        )?;
-    }
-    Ok(())
-}
-
-fn insert_current_token_v2_metadatas(
-    conn: &mut PgConnection,
-    items_to_insert: &[CurrentTokenV2Metadata],
-) -> Result<(), diesel::result::Error> {
-    use schema::current_token_v2_metadata::dsl::*;
-
-    let chunks = get_chunks(items_to_insert.len(), CurrentTokenV2Metadata::field_count());
-
-    for (start_ind, end_ind) in chunks {
-        execute_with_better_error(
-            conn,
-            diesel::insert_into(schema::current_token_v2_metadata::table)
-                .values(&items_to_insert[start_ind..end_ind])
-                .on_conflict((object_address, resource_type))
-                .do_update()
-                .set((
-                    data.eq(excluded(data)),
-                    state_key_hash.eq(excluded(state_key_hash)),
-                    last_transaction_version.eq(excluded(last_transaction_version)),
-                    inserted_at.eq(excluded(inserted_at)),
-                )),
-            Some(" WHERE current_token_v2_metadata.last_transaction_version <= excluded.last_transaction_version "),
-        )?;
-    }
-    Ok(())
-}
-
 #[async_trait]
 impl ProcessorTrait for TokenTransactionProcessor {
     fn name(&self) -> &'static str {
@@ -943,17 +617,6 @@ impl ProcessorTrait for TokenTransactionProcessor {
         let mut all_current_ans_lookups = all_current_ans_lookups
             .into_values()
             .collect::<Vec<CurrentAnsLookup>>();
-        // Token V2 processing which includes token v1
-        let (
-            collections_v2,
-            token_datas_v2,
-            token_ownerships_v2,
-            current_collections_v2,
-            current_token_ownerships_v2,
-            current_token_datas_v2,
-            token_activities_v2,
-            current_token_v2_metadata,
-        ) = parse_v2_token(&transactions, &table_handle_to_owner, &mut conn);
         all_current_ans_lookups
             .sort_by(|a, b| a.domain.cmp(&b.domain).then(a.subdomain.cmp(&b.subdomain)));
 
@@ -977,17 +640,6 @@ impl ProcessorTrait for TokenTransactionProcessor {
             all_current_token_claims,
             all_current_ans_lookups,
             all_nft_points,
-            // Token V2 stuff which will token v1 tables above
-            (
-                collections_v2,
-                token_datas_v2,
-                token_ownerships_v2,
-                current_collections_v2,
-                current_token_ownerships_v2,
-                current_token_datas_v2,
-                token_activities_v2,
-                current_token_v2_metadata,
-            ),
         );
         match tx_result {
             Ok(_) => Ok((start_version, end_version)),
@@ -1007,535 +659,4 @@ impl ProcessorTrait for TokenTransactionProcessor {
     fn connection_pool(&self) -> &PgDbPool {
         &self.connection_pool
     }
-}
-
-fn parse_v2_token(
-    transactions: &[Transaction],
-    table_handle_to_owner: &TableHandleToOwner,
-    conn: &mut PgPoolConnection,
-) -> (
-    Vec<CollectionV2>,
-    Vec<TokenDataV2>,
-    Vec<TokenOwnershipV2>,
-    Vec<CurrentCollectionV2>,
-    Vec<CurrentTokenDataV2>,
-    Vec<CurrentTokenOwnershipV2>,
-    Vec<TokenActivityV2>,
-    Vec<CurrentTokenV2Metadata>,
-) {
-    // Token V2 and V1 combined
-    let mut collections_v2 = vec![];
-    let mut token_datas_v2 = vec![];
-    let mut token_ownerships_v2 = vec![];
-    let mut token_activities_v2 = vec![];
-    let mut current_collections_v2: HashMap<CurrentCollectionV2PK, CurrentCollectionV2> =
-        HashMap::new();
-    let mut current_token_datas_v2: HashMap<CurrentTokenDataV2PK, CurrentTokenDataV2> =
-        HashMap::new();
-    let mut current_token_ownerships_v2: HashMap<
-        CurrentTokenOwnershipV2PK,
-        CurrentTokenOwnershipV2,
-    > = HashMap::new();
-    // Tracks prior ownership in case a token gets burned
-    let mut prior_nft_ownership: HashMap<String, NFTOwnershipV2> = HashMap::new();
-    // Get Metadata for token v2 by object
-    // We want to persist this through the entire batch so that even if a token is burned,
-    // we can still get the object core metadata for it
-    let mut token_v2_metadata_helper: TokenV2AggregatedDataMapping = HashMap::new();
-    // Basically token properties
-    let mut current_token_v2_metadata: HashMap<CurrentTokenV2MetadataPK, CurrentTokenV2Metadata> =
-        HashMap::new();
-
-    // Code above is inefficient (multiple passthroughs) so I'm approaching TokenV2 with a cleaner code structure
-    for txn in transactions {
-        let txn_data = txn.txn_data.as_ref().expect("Txn Data doesn't exit!");
-        let txn_version = txn.version as i64;
-        let txn_timestamp = parse_timestamp(txn.timestamp.as_ref().unwrap(), txn_version);
-        let transaction_info = txn.info.as_ref().expect("Transaction info doesn't exist!");
-
-        if let TxnData::User(user_txn) = txn_data {
-            let user_request = user_txn
-                .request
-                .as_ref()
-                .expect("Sends is not present in user txn");
-            let entry_function_id_str = get_entry_function_from_user_request(user_request);
-
-            // Get burn events for token v2 by object
-            let mut tokens_burned: TokenV2Burned = HashSet::new();
-
-            // Need to do a first pass to get all the objects
-            for (_, wsc) in transaction_info.changes.iter().enumerate() {
-                if let Change::WriteResource(wr) = wsc.change.as_ref().unwrap() {
-                    if let Some(object) =
-                        ObjectWithMetadata::from_write_resource(wr, txn_version).unwrap()
-                    {
-                        token_v2_metadata_helper.insert(
-                            standardize_address(&wr.address.to_string()),
-                            TokenV2AggregatedData {
-                                aptos_collection: None,
-                                fixed_supply: None,
-                                object,
-                                unlimited_supply: None,
-                                property_map: None,
-                                transfer_event: None,
-                                token: None,
-                                fungible_asset_metadata: None,
-                                fungible_asset_supply: None,
-                                fungible_asset_store: None,
-                            },
-                        );
-                    }
-                }
-            }
-
-            // Need to do a second pass to get all the structs related to the object
-            for (_, wsc) in transaction_info.changes.iter().enumerate() {
-                if let Change::WriteResource(wr) = wsc.change.as_ref().unwrap() {
-                    let address = standardize_address(&wr.address.to_string());
-                    if let Some(aggregated_data) = token_v2_metadata_helper.get_mut(&address) {
-                        if let Some(fixed_supply) =
-                            FixedSupply::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.fixed_supply = Some(fixed_supply);
-                        }
-                        if let Some(unlimited_supply) =
-                            UnlimitedSupply::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.unlimited_supply = Some(unlimited_supply);
-                        }
-                        if let Some(aptos_collection) =
-                            AptosCollection::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.aptos_collection = Some(aptos_collection);
-                        }
-                        if let Some(property_map) =
-                            PropertyMapModel::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.property_map = Some(property_map);
-                        }
-                        if let Some(token) = TokenV2::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.token = Some(token);
-                        }
-                        if let Some(fungible_asset_metadata) =
-                            FungibleAssetMetadata::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.fungible_asset_metadata = Some(fungible_asset_metadata);
-                        }
-                        if let Some(fungible_asset_supply) =
-                            FungibleAssetSupply::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.fungible_asset_supply = Some(fungible_asset_supply);
-                        }
-                        if let Some(fungible_asset_store) =
-                            FungibleAssetStore::from_write_resource(wr, txn_version).unwrap()
-                        {
-                            aggregated_data.fungible_asset_store = Some(fungible_asset_store);
-                        }
-                    }
-                }
-            }
-
-            // Pass through events to get the burn events and token activities v2
-            // This needs to be here because we need the metadata above for token activities
-            // and burn / transfer events need to come before the next section
-            for (index, event) in user_txn.events.iter().enumerate() {
-                if let Some(burn_event) = BurnEvent::from_event(event, txn_version).unwrap() {
-                    tokens_burned.insert(burn_event.get_token_address());
-                }
-                if let Some(transfer_event) = TransferEvent::from_event(event, txn_version).unwrap()
-                {
-                    if let Some(aggregated_data) =
-                        token_v2_metadata_helper.get_mut(&transfer_event.get_object_address())
-                    {
-                        // we don't want index to be 0 otherwise we might have collision with write set change index
-                        let index = if index == 0 {
-                            user_txn.events.len()
-                        } else {
-                            index
-                        };
-                        aggregated_data.transfer_event = Some((index as i64, transfer_event));
-                    }
-                }
-                // handling all the token v1 events
-                if let Some(event) = TokenActivityV2::get_v1_from_parsed_event(
-                    event,
-                    txn_version,
-                    txn_timestamp,
-                    index as i64,
-                    &entry_function_id_str,
-                )
-                .unwrap()
-                {
-                    token_activities_v2.push(event);
-                }
-                // handling all the token v2 events
-                if let Some(event) = TokenActivityV2::get_nft_v2_from_parsed_event(
-                    event,
-                    txn_version,
-                    txn_timestamp,
-                    index as i64,
-                    &entry_function_id_str,
-                    &token_v2_metadata_helper,
-                )
-                .unwrap()
-                {
-                    token_activities_v2.push(event);
-                }
-                // handling all the token v2 events
-                if let Some(event) = TokenActivityV2::get_ft_v2_from_parsed_event(
-                    event,
-                    txn_version,
-                    txn_timestamp,
-                    index as i64,
-                    &entry_function_id_str,
-                    &token_v2_metadata_helper,
-                    conn,
-                )
-                .unwrap()
-                {
-                    token_activities_v2.push(event);
-                }
-            }
-
-            for (index, wsc) in transaction_info.changes.iter().enumerate() {
-                let wsc_index = index as i64;
-                match wsc.change.as_ref().unwrap() {
-                    Change::WriteTableItem(table_item) => {
-                        if let Some((collection, current_collection)) =
-                            CollectionV2::get_v1_from_write_table_item(
-                                table_item,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                table_handle_to_owner,
-                                conn,
-                            )
-                            .unwrap()
-                        {
-                            collections_v2.push(collection);
-                            current_collections_v2.insert(
-                                current_collection.collection_id.clone(),
-                                current_collection,
-                            );
-                        }
-                        if let Some((token_data, current_token_data)) =
-                            TokenDataV2::get_v1_from_write_table_item(
-                                table_item,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                            )
-                            .unwrap()
-                        {
-                            token_datas_v2.push(token_data);
-                            current_token_datas_v2.insert(
-                                current_token_data.token_data_id.clone(),
-                                current_token_data,
-                            );
-                        }
-                        if let Some((token_ownership, current_token_ownership)) =
-                            TokenOwnershipV2::get_v1_from_write_table_item(
-                                table_item,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                table_handle_to_owner,
-                            )
-                            .unwrap()
-                        {
-                            token_ownerships_v2.push(token_ownership);
-                            if let Some(cto) = current_token_ownership {
-                                prior_nft_ownership.insert(
-                                    cto.token_data_id.clone(),
-                                    NFTOwnershipV2 {
-                                        token_data_id: cto.token_data_id.clone(),
-                                        owner_address: cto.owner_address.clone(),
-                                        is_soulbound: cto.is_soulbound_v2,
-                                    },
-                                );
-                                current_token_ownerships_v2.insert(
-                                    (
-                                        cto.token_data_id.clone(),
-                                        cto.property_version_v1.clone(),
-                                        cto.owner_address.clone(),
-                                        cto.storage_id.clone(),
-                                    ),
-                                    cto,
-                                );
-                            }
-                        }
-                    },
-                    Change::DeleteTableItem(table_item) => {
-                        if let Some((token_ownership, current_token_ownership)) =
-                            TokenOwnershipV2::get_v1_from_delete_table_item(
-                                table_item,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                table_handle_to_owner,
-                            )
-                            .unwrap()
-                        {
-                            token_ownerships_v2.push(token_ownership);
-                            if let Some(cto) = current_token_ownership {
-                                prior_nft_ownership.insert(
-                                    cto.token_data_id.clone(),
-                                    NFTOwnershipV2 {
-                                        token_data_id: cto.token_data_id.clone(),
-                                        owner_address: cto.owner_address.clone(),
-                                        is_soulbound: cto.is_soulbound_v2,
-                                    },
-                                );
-                                current_token_ownerships_v2.insert(
-                                    (
-                                        cto.token_data_id.clone(),
-                                        cto.property_version_v1.clone(),
-                                        cto.owner_address.clone(),
-                                        cto.storage_id.clone(),
-                                    ),
-                                    cto,
-                                );
-                            }
-                        }
-                    },
-                    Change::WriteResource(resource) => {
-                        if let Some((collection, current_collection)) =
-                            CollectionV2::get_v2_from_write_resource(
-                                resource,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                &token_v2_metadata_helper,
-                            )
-                            .unwrap()
-                        {
-                            collections_v2.push(collection);
-                            current_collections_v2.insert(
-                                current_collection.collection_id.clone(),
-                                current_collection,
-                            );
-                        }
-                        if let Some((token_data, current_token_data)) =
-                            TokenDataV2::get_v2_from_write_resource(
-                                resource,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                &token_v2_metadata_helper,
-                            )
-                            .unwrap()
-                        {
-                            // Add NFT ownership
-                            if let Some(inner) = TokenOwnershipV2::get_nft_v2_from_token_data(
-                                &token_data,
-                                &token_v2_metadata_helper,
-                            )
-                            .unwrap()
-                            {
-                                let (
-                                    nft_ownership,
-                                    current_nft_ownership,
-                                    from_nft_ownership,
-                                    from_current_nft_ownership,
-                                ) = inner;
-                                token_ownerships_v2.push(nft_ownership);
-                                // this is used to persist latest owner for burn event handling
-                                prior_nft_ownership.insert(
-                                    current_nft_ownership.token_data_id.clone(),
-                                    NFTOwnershipV2 {
-                                        token_data_id: current_nft_ownership.token_data_id.clone(),
-                                        owner_address: current_nft_ownership.owner_address.clone(),
-                                        is_soulbound: current_nft_ownership.is_soulbound_v2,
-                                    },
-                                );
-                                current_token_ownerships_v2.insert(
-                                    (
-                                        current_nft_ownership.token_data_id.clone(),
-                                        current_nft_ownership.property_version_v1.clone(),
-                                        current_nft_ownership.owner_address.clone(),
-                                        current_nft_ownership.storage_id.clone(),
-                                    ),
-                                    current_nft_ownership,
-                                );
-                                // Add the previous owner of the token transfer
-                                if let Some(from_nft_ownership) = from_nft_ownership {
-                                    let from_current_nft_ownership =
-                                        from_current_nft_ownership.unwrap();
-                                    token_ownerships_v2.push(from_nft_ownership);
-                                    current_token_ownerships_v2.insert(
-                                        (
-                                            from_current_nft_ownership.token_data_id.clone(),
-                                            from_current_nft_ownership.property_version_v1.clone(),
-                                            from_current_nft_ownership.owner_address.clone(),
-                                            from_current_nft_ownership.storage_id.clone(),
-                                        ),
-                                        from_current_nft_ownership,
-                                    );
-                                }
-                            }
-                            token_datas_v2.push(token_data);
-                            current_token_datas_v2.insert(
-                                current_token_data.token_data_id.clone(),
-                                current_token_data,
-                            );
-                        }
-
-                        // Add burned NFT handling
-                        if let Some((nft_ownership, current_nft_ownership)) =
-                            TokenOwnershipV2::get_burned_nft_v2_from_write_resource(
-                                resource,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                &tokens_burned,
-                            )
-                            .unwrap()
-                        {
-                            token_ownerships_v2.push(nft_ownership);
-                            prior_nft_ownership.insert(
-                                current_nft_ownership.token_data_id.clone(),
-                                NFTOwnershipV2 {
-                                    token_data_id: current_nft_ownership.token_data_id.clone(),
-                                    owner_address: current_nft_ownership.owner_address.clone(),
-                                    is_soulbound: current_nft_ownership.is_soulbound_v2,
-                                },
-                            );
-                            current_token_ownerships_v2.insert(
-                                (
-                                    current_nft_ownership.token_data_id.clone(),
-                                    current_nft_ownership.property_version_v1.clone(),
-                                    current_nft_ownership.owner_address.clone(),
-                                    current_nft_ownership.storage_id.clone(),
-                                ),
-                                current_nft_ownership,
-                            );
-                        }
-
-                        // Add fungible token handling
-                        if let Some((ft_ownership, current_ft_ownership)) =
-                            TokenOwnershipV2::get_ft_v2_from_write_resource(
-                                resource,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                &token_v2_metadata_helper,
-                                conn,
-                            )
-                            .unwrap()
-                        {
-                            token_ownerships_v2.push(ft_ownership);
-                            current_token_ownerships_v2.insert(
-                                (
-                                    current_ft_ownership.token_data_id.clone(),
-                                    current_ft_ownership.property_version_v1.clone(),
-                                    current_ft_ownership.owner_address.clone(),
-                                    current_ft_ownership.storage_id.clone(),
-                                ),
-                                current_ft_ownership,
-                            );
-                        }
-
-                        // Track token properties
-                        if let Some(token_metadata) = CurrentTokenV2Metadata::from_write_resource(
-                            resource,
-                            txn_version,
-                            &token_v2_metadata_helper,
-                        )
-                        .unwrap()
-                        {
-                            current_token_v2_metadata.insert(
-                                (
-                                    token_metadata.object_address.clone(),
-                                    token_metadata.resource_type.clone(),
-                                ),
-                                token_metadata,
-                            );
-                        }
-                    },
-                    Change::DeleteResource(resource) => {
-                        // Add burned NFT handling
-                        if let Some((nft_ownership, current_nft_ownership)) =
-                            TokenOwnershipV2::get_burned_nft_v2_from_delete_resource(
-                                resource,
-                                txn_version,
-                                wsc_index,
-                                txn_timestamp,
-                                &prior_nft_ownership,
-                                &tokens_burned,
-                                conn,
-                            )
-                            .unwrap()
-                        {
-                            token_ownerships_v2.push(nft_ownership);
-                            prior_nft_ownership.insert(
-                                current_nft_ownership.token_data_id.clone(),
-                                NFTOwnershipV2 {
-                                    token_data_id: current_nft_ownership.token_data_id.clone(),
-                                    owner_address: current_nft_ownership.owner_address.clone(),
-                                    is_soulbound: current_nft_ownership.is_soulbound_v2,
-                                },
-                            );
-                            current_token_ownerships_v2.insert(
-                                (
-                                    current_nft_ownership.token_data_id.clone(),
-                                    current_nft_ownership.property_version_v1.clone(),
-                                    current_nft_ownership.owner_address.clone(),
-                                    current_nft_ownership.storage_id.clone(),
-                                ),
-                                current_nft_ownership,
-                            );
-                        }
-                    },
-                    _ => {},
-                }
-            }
-        }
-    }
-
-    // Getting list of values and sorting by pk in order to avoid postgres deadlock since we're doing multi threaded db writes
-    let mut current_collections_v2 = current_collections_v2
-        .into_values()
-        .collect::<Vec<CurrentCollectionV2>>();
-    let mut current_token_datas_v2 = current_token_datas_v2
-        .into_values()
-        .collect::<Vec<CurrentTokenDataV2>>();
-    let mut current_token_ownerships_v2 = current_token_ownerships_v2
-        .into_values()
-        .collect::<Vec<CurrentTokenOwnershipV2>>();
-    let mut current_token_v2_metadata = current_token_v2_metadata
-        .into_values()
-        .collect::<Vec<CurrentTokenV2Metadata>>();
-
-    // Sort by PK
-    current_collections_v2.sort_by(|a, b| a.collection_id.cmp(&b.collection_id));
-    current_token_datas_v2.sort_by(|a, b| a.token_data_id.cmp(&b.token_data_id));
-    current_token_ownerships_v2.sort_by(|a, b| {
-        (
-            &a.token_data_id,
-            &a.property_version_v1,
-            &a.owner_address,
-            &a.storage_id,
-        )
-            .cmp(&(
-                &b.token_data_id,
-                &b.property_version_v1,
-                &b.owner_address,
-                &b.storage_id,
-            ))
-    });
-    current_token_v2_metadata.sort_by(|a, b| {
-        (&a.object_address, &a.resource_type).cmp(&(&b.object_address, &b.resource_type))
-    });
-
-    (
-        collections_v2,
-        token_datas_v2,
-        token_ownerships_v2,
-        current_collections_v2,
-        current_token_datas_v2,
-        current_token_ownerships_v2,
-        token_activities_v2,
-        current_token_v2_metadata,
-    )
 }
